Only unlink if report exists
For bill reports (the only ones not in html), if we cancel the dialog, the
tempfile is not even created, so there's nothing to unlink.

+++ stoqlib/gui/utils/printing.py
@@ -220,6 +220,8 @@ class PrintOperationPoppler(PrintOperation):
         page.render_for_printing(cr)
 
     def done(self):
+        if not os.path.isfile(self._report.filename):
+            return
         os.unlink(self._report.filename)
